misc: Move BillingService under Subscription namespace
